utat data download

+++ frontend/pages/api/utah/create.js
@@ -4,9 +4,9 @@ export default async function handler(req, res) {
   if (req.method === "POST") {
     try {
       const form = { id: req.body.id, tag: req.body.tag };
-      form.number_one = parseInt(req.body.number_one);
-      if (req.body.number_two) form.number_two = parseInt(req.body.number_two);
-      if (req.body.number_three) form.number_three = parseInt(req.body.number_three);
+      form.number_one = parseFloat(req.body.number_one);
+      if (req.body.number_two) form.number_two = parseFloat(req.body.number_two);
+      if (req.body.number_three) form.number_three = parseFloat(req.body.number_three);
       await new DynamoDbManager("Primaryhealth-utah").create(form);
       res.status(200).json({message: 'success'})
     } catch {
